+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "spec/common/ably-common"]
+	path = spec/common/ably-common
+	url = git@github.com:ably/ably-common.git
