+++ spec/rest/capability.test.js
@@ -81,7 +81,7 @@ define(['ably', 'shared_helper'], function(Ably, helper) {
 	exports.authcapability2 = function(test) {
 		test.expect(1);
 	var testKeyOpts = {key: testApp.keys[1].keyStr};
-		var testCapability = {testchannel:['subscribe']};
+		var testCapability = {"canpublish:test":['subscribe']};
 		rest.auth.requestToken(testKeyOpts, {capability: testCapability}, function(err, tokenDetails) {
 			if(err) {
 				test.equal(err.statusCode, 401, 'Verify request rejected with insufficient capability');
@@ -98,7 +98,7 @@ define(['ably', 'shared_helper'], function(Ably, helper) {
 	 */
 	exports.authcapability3 = function(test) {
 		test.expect(1);
-	var testKeyOpts = {key: testApp.keys[4].keyStr};
+	var testKeyOpts = {key: testApp.keys[2].keyStr};
 		var testCapability = {channelx:['publish']};
 		rest.auth.requestToken(testKeyOpts, {capability: testCapability}, function(err, tokenDetails) {
 			if(err) {
@@ -116,7 +116,7 @@ define(['ably', 'shared_helper'], function(Ably, helper) {
 	 */
 	exports.authcapability4 = function(test) {
 		test.expect(1);
-	var testKeyOpts = {key: testApp.keys[4].keyStr};
+	var testKeyOpts = {key: testApp.keys[2].keyStr};
 		var testCapability = {channel2:['presence', 'subscribe']};
 		var expectedIntersection = {channel2:['subscribe']};
 		rest.auth.requestToken(testKeyOpts, {capability: testCapability}, function(err, tokenDetails) {
@@ -135,7 +135,7 @@ define(['ably', 'shared_helper'], function(Ably, helper) {
 	 */
 	exports.authcapability5 = function(test) {
 		test.expect(1);
-	var testKeyOpts = {key: testApp.keys[4].keyStr};
+	var testKeyOpts = {key: testApp.keys[2].keyStr};
 		var testCapability = {
 			channel2:['presence', 'subscribe'],
 			channelx:['presence', 'subscribe']
@@ -157,7 +157,7 @@ define(['ably', 'shared_helper'], function(Ably, helper) {
 	 */
 	exports.authcapability6 = function(test) {
 		test.expect(1);
-	var testKeyOpts = {key: testApp.keys[4].keyStr};
+	var testKeyOpts = {key: testApp.keys[2].keyStr};
 		var testCapability = {channel2:['*']};
 		var expectedIntersection = {channel2:['publish', 'subscribe']};
 		rest.auth.requestToken(testKeyOpts, {capability: testCapability}, function(err, tokenDetails) {
@@ -172,7 +172,7 @@ define(['ably', 'shared_helper'], function(Ably, helper) {
 	};
 	exports.authcapability7 = function(test) {
 		test.expect(1);
-	var testKeyOpts = {key: testApp.keys[4].keyStr};
+	var testKeyOpts = {key: testApp.keys[2].keyStr};
 		var testCapability = {channel6:['publish', 'subscribe']};
 		var expectedIntersection = {channel6:['publish', 'subscribe']};
 		rest.auth.requestToken(testKeyOpts, {capability: testCapability}, function(err, tokenDetails) {
@@ -191,7 +191,7 @@ define(['ably', 'shared_helper'], function(Ably, helper) {
 	 */
 	exports.authcapability8 = function(test) {
 		test.expect(1);
-	var testKeyOpts = {key: testApp.keys[2].keyStr};
+	var testKeyOpts = {key: testApp.keys[3].keyStr};
 		var testCapability = {cansubscribe:['subscribe']};
 		var expectedIntersection = testCapability;
 		rest.auth.requestToken(testKeyOpts, {capability: testCapability}, function(err, tokenDetails) {
@@ -206,7 +206,7 @@ define(['ably', 'shared_helper'], function(Ably, helper) {
 	};
 	exports.authcapability9 = function(test) {
 		test.expect(1);
-	var testKeyOpts = {key: testApp.keys[2].keyStr};
+	var testKeyOpts = {key: testApp.keys[1].keyStr};
 		var testCapability = {'canpublish:check':['publish']};
 		var expectedIntersection = testCapability;
 		rest.auth.requestToken(testKeyOpts, {capability: testCapability}, function(err, tokenDetails) {
@@ -221,7 +221,7 @@ define(['ably', 'shared_helper'], function(Ably, helper) {
 	};
 	exports.authcapability10 = function(test) {
 		test.expect(1);
-	var testKeyOpts = {key: testApp.keys[2].keyStr};
+	var testKeyOpts = {key: testApp.keys[3].keyStr};
 		var testCapability = {'cansubscribe:*':['subscribe']};
 		var expectedIntersection = testCapability;
 		rest.auth.requestToken(testKeyOpts, {capability: testCapability}, function(err, tokenDetails) {
